requirements.txt
Django==1.10
Pillow==4.4.0
django-widget-tweaks==1.4.1
django-watson
